refactor(data-model): the walk belongs to the act it runs in
`encodeValue()`, `decodeTagged()` and the format's own account of how a
container is written down move from `BaseCodecEngine` onto `EncodeAct` and
`DecodeAct`. **The threaded act parameter is gone**: the state one call
carries and the methods that consult it are now the same object, so there is
nothing left to thread.

An engine is what remains: its configuration -- the registry and the leniency
setting -- the two public entry points, and the factories that say which act
classes a format uses. `CodecEngineConfig` grows `registry`, which is the
growth it was introduced to absorb, and `registry` becomes public so an act
can read it.

## Why the hooks move too, rather than staying put

A hook reached from an act would not compile. They are `protected` on the
engine, and an act is not in the engine's hierarchy, so the call raises
`TS2445`. Moving the walk without them would have meant making a format's
internal contract public API. Moving them alongside dissolves the problem
instead: the walk and the hooks land in one hierarchy, where an ordinary
protected call reaches them.

## The JSON format, split three ways

`JsonCodecEngine` was 701 lines carrying both directions. Its encode half is
now `JsonEncodeAct` and its decode half `JsonDecodeAct`, with `wire-text.ts`
holding the three helpers both an engine and an act need -- `parseWireText()`,
`seemsLikeEncoded()` and `isEncodedInstance()` -- so that neither imports the
other and no cycle is possible.

Note the byte and string boundaries are NOT interchangeable and stay separate:
`decodeFromBytes()` parses raw text where `encodedFromSerializedForm()` checks
and strips the format prefix first.

Neither act class is package surface. `codec-json`'s barrel exports the engine
and the registry factory; the acts and this format's three own codecs are
reached internally, their unit tests importing them by direct path.

## Verification

Ablation, since a green suite proves little after a move this size. Each of
the moved decisions was broken in turn and every one turned the suite red: the
self-representing arm, the container dispatch, the terminal/nonterminal split,
the unknown-tag arm and the tag syntax check.

That found a real gap, and it was pre-existing rather than introduced here:
nothing pinned the cycle guard around a TAGGED value on the encode side.
Removing the `enter()` left the suite green. There is now a test for it -- a
nonterminal codec whose state names the value being encoded, closing a cycle
that no container arm reaches -- and removing either the `enter()` or the
`leave()` turns the suite red.

- `data-model`: `ok | 54 passed (2459 steps) | 0 failed`, against 2458 before,
  the one added step being that test.
- `it()` and `describe()` counts diffed against the base and unchanged.
- Gates green: `lint`, `check`, `fmt --check`, `check-docs`,
  `check-skill-facts`, `check-package-cycles`, `check-no-waitfor`,
  `check-conflict-markers`, `check-unused-deps`, `check-single-copy-deps`.

The formal spec's account of the engine's shape is updated: what an engine
supplies, what an act supplies, and why the walks belong to an act.
