+++ packages/data-model/src/codec-common/CodecEngineConfig.ts
@@ -1,3 +1,5 @@
+import type { CodecRegistry } from "./CodecRegistry.ts";
+
 /**
  * What an act of encoding or decoding reads off the engine that minted it:
  * the settings an engine holds for its whole lifetime, as against the state
@@ -8,15 +10,17 @@
  * engine's whole surface, and it keeps the two modules independent of each
  * other: an act never imports an engine. `BaseCodecEngine` satisfies this
  * interface, so an engine hands over itself.
- *
- * Expect this to grow as more of the walk moves onto the acts. Growing it is
- * the intended way to give an act something new to read, in preference to
- * widening what an act is handed.
  */
-export interface CodecEngineConfig {
+export interface CodecEngineConfig<Encoded> {
   /**
    * Whether a decode that fails produces a `ProblematicValue` instead of
    * throwing.
    */
   readonly lenient: boolean;
+
+  /**
+   * The codecs this engine encodes and decodes with, and so which classes it
+   * can carry.
+   */
+  readonly registry: CodecRegistry<Encoded>;
 }
