+++ packages/data-model/src/codec-common/DecodeAct.ts
@@ -1,6 +1,15 @@
+import { backtickQuote } from "@commonfabric/utils/markdown";
+
 import { deepFreeze } from "@/deep-freeze.ts";
-import type { FabricValue } from "@/interface.ts";
+import { BaseTerminalCodec } from "@/codec-interface/BaseTerminalCodec.ts";
+import type {
+  NonterminalCodec,
+  TerminalCodec,
+} from "@/codec-interface/interface.ts";
 import { toCompactDebugString } from "@/value-debug.ts";
+import { isCodecTypeTag } from "./isCodecTypeTag.ts";
+import { UnknownValue } from "./UnknownValue.ts";
+import type { FabricValue } from "@/interface.ts";
 import { BaseCodecAct } from "./BaseCodecAct.ts";
 import { ProblematicStateError } from "./ProblematicStateError.ts";
 import { ProblematicValue } from "./ProblematicValue.ts";
@@ -17,7 +26,48 @@ import { ProblematicValue } from "./ProblematicValue.ts";
  * settle is a property of the act: they run while a walk is in progress, and
  * every one of them is reached with the act already in hand.
  */
-export class DecodeAct extends BaseCodecAct {
+export abstract class DecodeAct<Encoded, SerializedForm = Encoded>
+  extends BaseCodecAct<Encoded> {
+  //
+  // Subclass contract
+  //
+
+  /**
+   * Converts this format's serialized form into the tree this act decodes,
+   * which is where a format checks that what it was handed is its own: a parse
+   * step, an envelope check, or nothing at all.
+   *
+   * What arrives is data off a channel and is not to be assumed well-formed. A
+   * form that is not this format's is refused by throwing
+   * `ProblematicStateError`, which the engine's `decode()` settles through
+   * {@link #settleThrown} -- so a lenient decode returns a `ProblematicValue`
+   * for a syntactic fault, exactly as it does for a fault found further in.
+   */
+  abstract encodedFromSerializedForm(data: SerializedForm): Encoded;
+
+  /**
+   * Decodes a transport tree back into fabric values.
+   *
+   * Whether cycles are guarded at all is the format's decision, made by
+   * whether this method enters a node: a format whose input it parses for
+   * itself cannot be handed a cycle, so it enters none and this act's set is
+   * never allocated.
+   *
+   * An implementation that does guard owes the act one thing: every object it
+   * is about to descend through goes through {@link #enterOrReport} first, and
+   * comes back out through {@link #leave} however the descent ends. That means
+   * the tagged form as much as a container -- a format whose transport can
+   * carry a graph can close a cycle through tagged nodes alone. Here rather
+   * than in {@link #decodeTagged}, because this method is the one that visits
+   * every node, and entering in both places would enter a state twice and
+   * report a cycle that is not there.
+   */
+  abstract decodeValue(data: Encoded): FabricValue;
+
+  //
+  // Instance members
+  //
+
   /**
    * Enters a node, and returns whether it was entered.
    *
@@ -149,5 +199,113 @@ export class DecodeAct extends BaseCodecAct {
       state,
       `object contains a key this runtime reserves: "${key}"`,
     );
+  }
+
+  /**
+   * Decodes one tagged value, dispatching on the tag through the registry. A
+   * subclass calls this once it has recognized a tagged form and taken off
+   * whatever meta-tags this format defines for itself.
+   *
+   * The tag's syntax is checked here rather than by each caller: `tag` is
+   * whatever a format found in tag position, of whatever type, and a subclass
+   * is not expected to know what a tag may look like.
+   *
+   * Frozen-ness contract: every value returned from here is deep-frozen, so
+   * callers do not each have to freeze, and a caller need not ask which arm
+   * produced what it was handed.
+   *
+   * Three of the arms below walk the state again -- a nonterminal codec's, an
+   * unknown tag's, and a malformed tag's -- and each carries `act` into that
+   * walk. Entering the state is not this method's business: it is
+   * {@link #decodeValue} that visits every node of the tree, the tagged form
+   * included, and entering there is what keeps one node from being entered
+   * twice.
+   */
+  decodeTagged(
+    tag: any,
+    rawState: Encoded,
+  ): FabricValue {
+    if (!isCodecTypeTag(tag)) {
+      // Anything that is not a tag syntactically is an encoding error whatever
+      // follows it, per Section 9 of the formal spec, and is reported rather
+      // than preserved as an `UnknownValue`: that form exists to round-trip a
+      // tag no codec claims, which presupposes a tag. Reported over the
+      // decoded state, so that a lenient result carries what arrived.
+      return this.reportMalformed(
+        tag,
+        this.decodeValue(rawState),
+        `tagged value has a malformed tag: ${
+          backtickQuote(toCompactDebugString(tag, 30))
+        }`,
+      );
+    }
+
+    const matched = this.registry.codecFromTag(tag);
+
+    if (matched === undefined) {
+      // A tag this registry does not carry, kept in the unknown form so that
+      // it round-trips.
+      return deepFreeze(new UnknownValue(tag, this.decodeValue(rawState)));
+    }
+
+    // A terminal codec takes the state exactly as it arrived; a nonterminal
+    // one takes it expanded. The casts restate what `instanceof` just
+    // established, which TypeScript drops on a generic class.
+    const terminal = matched instanceof BaseTerminalCodec;
+    const state = terminal ? rawState : this.decodeValue(rawState);
+
+    let decoded: FabricValue;
+
+    try {
+      decoded = terminal
+        ? (matched as TerminalCodec<Encoded>).decode(tag, rawState, this.env)
+        : (matched as NonterminalCodec).decode(
+          tag,
+          state as FabricValue,
+          this.env,
+        );
+    } catch (e: any) {
+      if (!this.config.lenient) {
+        // Normalized rather than rethrown: what a codec throws is not
+        // guaranteed to be an `Error`, let alone one naming the state it
+        // choked on. `fromThrown()` returns one that does -- the thrown
+        // value itself where it already qualifies, and otherwise a fresh one
+        // holding it as `cause`.
+        throw ProblematicStateError.fromThrown(tag, state, e);
+      }
+
+      // Report over the state the codec was actually handed, so that it says
+      // what the codec choked on.
+      return this.reportMalformed(
+        tag,
+        state,
+        e instanceof Error ? e.message : String(e),
+      );
+    }
+
+    if (
+      !this.config.lenient && (decoded instanceof ProblematicValue) &&
+      (matched.uniqueHandledClass !== ProblematicValue)
+    ) {
+      // The two ways a codec reports a state it will not accept -- throwing,
+      // and returning one of these -- are the codec author's choice and say
+      // nothing about what a caller wants. `lenient` is what says that, so
+      // this instance settles both into the same answer: a strict decode
+      // fails, whichever way the codec reported it.
+      //
+      // `ProblematicValue`'s own codec is exempt, because for that one a
+      // `ProblematicValue` is the successful product rather than a refusal.
+      // A payload under `Problematic@1` is a well-formed record of a past
+      // failure, and reading one is not a failure of this decode; without the
+      // exemption a strict reader could never read such a record back, which
+      // is most of what preserving one is for.
+      throw new ProblematicStateError(tag, decoded.state, decoded.error);
+    }
+
+    // A codec's `decode()` promises deep-frozen results rather than relying on
+    // every caller to freeze. That covers the codec's own product -- a
+    // `FabricPrimitive` is already frozen, making it an O(1) cache hit -- and
+    // the lenient fallback above alike.
+    return deepFreeze(decoded);
   }
 }
