+++ packages/data-model/src/codec-json/JsonCodecEngine.ts
@@ -1,23 +1,18 @@
 import { backtickQuote } from "@commonfabric/utils/markdown";
-import { isPlainObject, isUnsafeObjectKey } from "@commonfabric/utils/types";
-import { utf8SortedKeysOf } from "@commonfabric/utils/utf8";
 
 import type { FabricValue } from "@/interface.ts";
 import { BaseCodecEngine } from "@/codec-common/BaseCodecEngine.ts";
-import { ProblematicStateError } from "@/codec-common/ProblematicStateError.ts";
-import { DecodeAct } from "@/codec-common/DecodeAct.ts";
-import { EncodeAct } from "@/codec-common/EncodeAct.ts";
+import { JsonDecodeAct } from "./JsonDecodeAct.ts";
+import { JsonEncodeAct } from "./JsonEncodeAct.ts";
+import { parseWireText, seemsLikeEncoded } from "./wire-text.ts";
 import { NULL_LIVE_ENVIRONMENT } from "@/codec-interface/NullLiveEnvironment.ts";
-import { toCompactDebugString } from "@/value-debug.ts";
 import { CODEC, type LiveEnvironment } from "@/codec-interface/interface.ts";
-import { deepFreeze } from "@/deep-freeze.ts";
 import { NullLiveEnvironment } from "@/codec-interface/NullLiveEnvironment.ts";
 import { UnknownValue } from "@/codec-common/UnknownValue.ts";
 import { ProblematicValue } from "@/codec-common/ProblematicValue.ts";
 import { ENCODING_PREFIX_TAG, type JsonCodecValue } from "./interface.ts";
 import { createBaseJsonRegistry } from "./createBaseJsonRegistry.ts";
 import type { CodecRegistry } from "@/codec-common/CodecRegistry.ts";
-import { CODEC_META_TAGS } from "@/codec-interface/codec-meta-tags.ts";
 
 /**
  * Whole-value JSON codec implementing the `/<Type>@<Version>` wire format from
@@ -55,8 +50,8 @@ export class JsonCodecEngine extends BaseCodecEngine<JsonCodecValue, string> {
   //
 
   /** @inheritDoc */
-  protected override newEncodeAct(env: LiveEnvironment): EncodeAct {
-    return new EncodeAct(this, env);
+  protected override newEncodeAct(env: LiveEnvironment): JsonEncodeAct {
+    return new JsonEncodeAct(this, env);
   }
 
   /**
@@ -69,43 +64,8 @@ export class JsonCodecEngine extends BaseCodecEngine<JsonCodecValue, string> {
   protected override newDecodeAct(
     env: LiveEnvironment,
     _data: string,
-  ): DecodeAct {
-    return new DecodeAct(this, env);
-  }
-
-  /**
-   * @inheritDoc
-   *
-   * Stringifies the walked tree and prefixes the format tag.
-   */
-  protected override serializedFromEncoded(
-    encoded: JsonCodecValue,
-    _act: EncodeAct,
-  ): string {
-    return ENCODING_PREFIX_TAG + JSON.stringify(encoded);
-  }
-
-  /**
-   * @inheritDoc
-   *
-   * Checks the format tag and parses what follows it. A string without the tag
-   * is not this format's serialized form at all, which is refused here rather
-   * than walked -- and settles against `lenient` like any other malformation
-   * off a channel.
-   */
-  protected override encodedFromSerializedForm(data: string): JsonCodecValue {
-    if (!JsonCodecEngine.seemsLikeEncoded(data)) {
-      const excerpt = (data.length <= 50) ? data : `${data.slice(0, 50)}...`;
-      throw new ProblematicStateError(
-        "",
-        excerpt,
-        `Not a JSON-encoded \`FabricValue\` string: ${backtickQuote(excerpt)}`,
-      );
-    }
-
-    return JsonCodecEngine.#parseWireText(
-      data.slice(ENCODING_PREFIX_TAG.length),
-    );
+  ): JsonDecodeAct {
+    return new JsonDecodeAct(this, env);
   }
 
   /** Encodes a fabric value to UTF-8 JSON bytes. */
@@ -113,9 +73,7 @@ export class JsonCodecEngine extends BaseCodecEngine<JsonCodecValue, string> {
     value: FabricValue,
     env: LiveEnvironment = NULL_LIVE_ENVIRONMENT,
   ): Uint8Array {
-    return JsonCodecEngine.#toBytes(
-      this.encodeValue(value, this.newEncodeAct(env)),
-    );
+    return JsonCodecEngine.#toBytes(this.newEncodeAct(env).encodeValue(value));
   }
 
   /**
@@ -132,284 +90,12 @@ export class JsonCodecEngine extends BaseCodecEngine<JsonCodecValue, string> {
     let tree: JsonCodecValue;
 
     try {
-      tree = JsonCodecEngine.#parseWireText(text);
+      tree = parseWireText(text);
     } catch (e) {
       return act.settleThrown(e);
     }
 
-    return this.decodeValue(tree, act);
-  }
-
-  /**
-   * @inheritDoc
-   *
-   * Prepends `/` to the tag to produce the JSON key. See Section 5.2 of the
-   * formal spec.
-   *
-   * The result is not frozen. An encode-side tree is stringified and
-   * discarded without ever reaching a caller, so the deep-frozen invariant
-   * `JsonCodecValue` states does not cover it, and freezing every tagged node
-   * on the way out is measurable on small values. The meta-tag call sites
-   * below freeze what they build, where the cost is already paid by the
-   * rebuild around it.
-   */
-  protected override wrapTag(
-    tag: string,
-    state: JsonCodecValue,
-    _act: EncodeAct,
-  ): JsonCodecValue {
-    return { [`/${tag}`]: state } as JsonCodecValue;
-  }
-
-  /**
-   * @inheritDoc
-   *
-   * A run of holes is represented as a `/hole` count, JSON having no way
-   * to write an absent index.
-   */
-  protected override encodeArray(
-    value: readonly FabricValue[],
-    act: EncodeAct,
-  ): JsonCodecValue {
-    act.enter(value);
-
-    const result: JsonCodecValue[] = [];
-    try {
-      let i = 0;
-      while (i < value.length) {
-        if (!(i in value)) {
-          let count = 0;
-          while (i < value.length && !(i in value)) {
-            count++;
-            i++;
-          }
-          result.push(
-            Object.freeze(this.wrapTag(CODEC_META_TAGS.hole, count, act)),
-          );
-        } else {
-          result.push(this.encodeValue(value[i]!, act));
-          i++;
-        }
-      }
-    } finally {
-      act.leave(value);
-    }
-
-    return result as JsonCodecValue;
-  }
-
-  /**
-   * @inheritDoc
-   *
-   * Keys are visited in UTF-8 byte order, matching the canonical order
-   * `value-hash.ts` uses, so that this encoding is deterministic across
-   * implementations and across objects whose keys differ only in insertion
-   * order. See `3-json-encoding.md` Section 10.
-   *
-   * A `/`-prefixed key collides with the tag form, so an object bearing one is
-   * escaped per Section 5.6: all values are encoded first, and if every one is
-   * quote-safe the whole object is wrapped in `/quote` with any `/quote`
-   * children collapsed into it, and otherwise in `/object` so that the decoder
-   * walks the entries.
-   */
-  protected override encodePlainObject(
-    value: Record<string, FabricValue>,
-    act: EncodeAct,
-  ): JsonCodecValue {
-    act.enter(value);
-
-    const result: Record<string, JsonCodecValue> = {};
-    let anySlashKey = false;
-    try {
-      for (const key of utf8SortedKeysOf(value)) {
-        JsonCodecEngine.assertEncodableKey(key);
-
-        if (key.startsWith("/")) {
-          anySlashKey = true;
-        }
-        result[key] = this.encodeValue(value[key]!, act);
-      }
-    } finally {
-      act.leave(value);
-    }
-
-    if (anySlashKey) {
-      if (Object.values(result).every((v) => JsonCodecEngine.#isQuoteSafe(v))) {
-        const unquoted = Object.freeze(
-          Object.fromEntries(
-            Object.entries(result).map((
-              [k, v],
-            ) => [k, JsonCodecEngine.#unquote(v)]),
-          ),
-        );
-        return Object.freeze(
-          this.wrapTag(CODEC_META_TAGS.quote, unquoted, act),
-        );
-      }
-      return Object.freeze(this.wrapTag(CODEC_META_TAGS.object, result, act));
-    }
-
-    return result as JsonCodecValue;
-  }
-
-  /**
-   * Decodes a codec-value tree back into fabric values. See Section 4.5 of
-   * the formal spec.
-   *
-   * Frozen-ness contract: values returned via the codec dispatch arm are
-   * guaranteed deep-frozen at this boundary, so callers do not each have to
-   * freeze. The unknown-tag fallback (`UnknownValue`) is a separate arm and is
-   * intentionally NOT covered by this contract.
-   */
-  protected override decodeValue(
-    data: JsonCodecValue,
-    act: DecodeAct,
-  ): FabricValue {
-    const decoded = JsonCodecEngine.#unwrapTag(data);
-    if (decoded !== null) {
-      const { tag, state: rawState } = decoded;
-
-      // `CODEC_META_TAGS.quote` literal handling (Section 5.6).
-      if (tag === CODEC_META_TAGS.quote) {
-        return rawState;
-      }
-
-      // `CODEC_META_TAGS.object` unwrapping (Section 5.6).
-      if (tag === CODEC_META_TAGS.object) {
-        const inner = rawState as Record<string, JsonCodecValue>;
-        const result: Record<string, FabricValue> = {};
-        for (const [key, val] of Object.entries(inner)) {
-          // Same reservation as the plain-object arm below: the assignment
-          // cannot rebuild these names.
-          if (isUnsafeObjectKey(key)) {
-            return act.reportReservedKey(key, inner);
-          }
-          result[key] = this.decodeValue(val, act);
-        }
-        return Object.freeze(result);
-      }
-
-      // `/quote` and `/object` returned above, so no codec ever sees their
-      // state, and `/quote` contents alone go undecoded.
-      return this.decodeTagged(tag, rawState, act);
-    }
-
-    // Primitives pass through.
-    if (
-      data === null || typeof data === "boolean" ||
-      typeof data === "number" || typeof data === "string"
-    ) {
-      return data;
-    }
-
-    if (Array.isArray(data)) {
-      return this.#decodeArray(data, act);
-    }
-
-    // `Array.isArray()` above removed the array arm, but TypeScript keeps it
-    // in the union; the remaining member is the record.
-    return this.#decodePlainObject(
-      data as Record<string, JsonCodecValue>,
-      act,
-    );
-  }
-
-  /**
-   * Arrays: recursively decode elements.
-   *
-   * One pass. A `/hole` run advances the write index past the indices it
-   * stands for, leaving them absent, and the final length is set from that
-   * index so that a run in the last position is preserved. Counting the
-   * logical length first would mean walking and unwrapping every entry a
-   * second time, for a number this pass arrives at anyway.
-   *
-   * The result is still sized up front, at the entry count. That is exact
-   * whenever the array has no holes, which is the ordinary case, and an
-   * underestimate otherwise -- growing from there beats growing from empty,
-   * and a short array of holes is common enough to be worth not pessimizing.
-   *
-   * A run's count is validated, wire data being untrusted. Left unchecked it
-   * is added to the write index directly, so a string concatenates onto it
-   * and a negative or fractional one makes the length assignment throw --
-   * failures with no bearing on what went wrong. A run stands for at least
-   * one absent index, and anything else is reported instead.
-   */
-  #decodeArray(
-    data: readonly JsonCodecValue[],
-    act: DecodeAct,
-  ): FabricValue {
-    const result: FabricValue[] = new Array(data.length);
-    let targetIndex = 0;
-    for (const entry of data) {
-      const entryDecoded = JsonCodecEngine.#unwrapTag(entry);
-      if (
-        entryDecoded !== null && entryDecoded.tag === CODEC_META_TAGS.hole
-      ) {
-        const count = entryDecoded.state;
-        if (!JsonCodecEngine.#isHoleCount(count)) {
-          return act.reportMalformed(
-            CODEC_META_TAGS.hole,
-            count,
-            `hole: expected a positive integer count, got ${
-              backtickQuote(toCompactDebugString(count, 30))
-            }`,
-          );
-        }
-        targetIndex += count;
-      } else {
-        result[targetIndex] = this.decodeValue(entry, act);
-        targetIndex++;
-      }
-    }
-
-    // The total is bounded here rather than each advance being bounded as
-    // it happens, because both a single run and the running total can pass
-    // what an array may hold, and one check at the end covers both. Beyond
-    // that, the assignment below throws `RangeError` from the array
-    // machinery, which says nothing about the wire that caused it.
-    const MAX_ARRAY_LENGTH = 0xffff_ffff;
-    if (targetIndex > MAX_ARRAY_LENGTH) {
-      return act.reportMalformed(
-        CODEC_META_TAGS.hole,
-        data,
-        `hole: runs total ${targetIndex} elements, past the ` +
-          `${MAX_ARRAY_LENGTH} an array can hold`,
-      );
-    }
-
-    result.length = targetIndex;
-    return Object.freeze(result);
-  }
-
-  /**
-   * Plain objects: recursively decode values and freeze. Any
-   * `/`-prefixed key is reserved per spec — return `ProblematicValue` on
-   * first occurrence rather than silently round-tripping the object.
-   */
-  #decodePlainObject(
-    data: Record<string, JsonCodecValue>,
-    act: DecodeAct,
-  ): FabricValue {
-    const result: Record<string, FabricValue> = {};
-    for (const [key, val] of Object.entries(data)) {
-      if (key.startsWith("/")) {
-        return act.reportMalformed(
-          key.slice(1),
-          data,
-          `object contains reserved /-prefixed key: "${key}"`,
-        );
-      }
-      // A name this runtime reserves cannot be rebuilt by the assignment
-      // below: `__proto__` would repoint the result's prototype instead of
-      // becoming a property. Such a record cannot have been written by this
-      // implementation, whose write path refuses it, so report it rather than
-      // decoding something the bytes do not say.
-      if (isUnsafeObjectKey(key)) {
-        return act.reportReservedKey(key, data);
-      }
-      result[key] = this.decodeValue(val, act);
-    }
-    return Object.freeze(result);
+    return act.decodeValue(tree);
   }
 
   //
@@ -465,7 +151,7 @@ export class JsonCodecEngine extends BaseCodecEngine<JsonCodecValue, string> {
    * tag.
    */
   static seemsLikeEncoded(value: string): boolean {
-    return value.startsWith(ENCODING_PREFIX_TAG);
+    return seemsLikeEncoded(value);
   }
 
   /**
@@ -578,124 +264,8 @@ export class JsonCodecEngine extends BaseCodecEngine<JsonCodecValue, string> {
     return encoded;
   }
 
-  /**
-   * Unwraps a wire representation. Detects single-key objects with `/`-prefixed
-   * keys. Returns `{ tag, state }` or `null` if not a tagged value. The
-   * returned `state` is extracted directly from `data`, so if `data` is
-   * deep-frozen (as it should be) then `state` will be too.
-   *
-   * See Section 5.4 of the formal spec.
-   */
-  static #unwrapTag(
-    data: JsonCodecValue,
-  ): { tag: string; state: JsonCodecValue } | null {
-    if (!isPlainObject(data)) {
-      return null;
-    }
-
-    if (!JsonCodecEngine.#isEncodedInstance(data)) {
-      return null;
-    }
-
-    // `#isEncodedInstance()` guaranteed a single-property object, so this
-    // destructures that one entry.
-    const [key, value] = Object.entries(data)[0]!;
-    return { tag: key.slice(1), state: value };
-  }
-
   /** Converts a codec-value tree to UTF-8-encoded JSON bytes. */
   static #toBytes(data: JsonCodecValue): Uint8Array {
     return JsonCodecEngine.#textEncoder.encode(JSON.stringify(data));
-  }
-
-  /**
-   * Indicates whether `count` is usable as a `/hole` run length: a safe
-   * integer of at least one. A run always stands for at least one absent
-   * index, so zero is refused along with everything else that is not a count.
-   */
-  static #isHoleCount(count: JsonCodecValue): count is number {
-    // `isSafeInteger()` returns `false` for a non-number, but it cannot be a
-    // TypeScript type predicate on `number` since it also returns `false` for
-    // plenty of numbers. So, the subsequent cast `as number` is safe by
-    // construction but is nonetheless required.
-    return Number.isSafeInteger(count) && ((count as number) >= 1);
-  }
-
-  /**
-   * Returns true if `v` is a single-key object whose key starts with `/` --
-   * the wire form of an encoded instance (tag-wrapped value).
-   */
-  static #isEncodedInstance(v: JsonCodecValue): boolean {
-    if (v === null || typeof v !== "object" || Array.isArray(v)) return false;
-    const keys = Object.keys(v);
-    return keys.length === 1 && keys[0]!.startsWith("/");
-  }
-
-  /**
-   * Returns true if the already-encoded codec value `v` can be embedded
-   * inside a /quote wrap without inner decoding: primitives, plain
-   * objects/arrays free of non-/quote encoded instances, and /quote-wrapped
-   * values (which `#unquote()` can collapse).
-   */
-  static #isQuoteSafe(v: JsonCodecValue): boolean {
-    if (v === null || typeof v !== "object") return true;
-    if (Array.isArray(v)) {
-      return v.every((item) => JsonCodecEngine.#isQuoteSafe(item));
-    }
-    if (!JsonCodecEngine.#isEncodedInstance(v)) {
-      return Object.values(v).every((item) =>
-        JsonCodecEngine.#isQuoteSafe(item as JsonCodecValue)
-      );
-    }
-    return Object.keys(v)[0] === "/quote";
-  }
-
-  /**
-   * Unwraps /quote forms one level so their literal content can be embedded
-   * directly inside a parent /quote. The inner content of a /quote is already
-   * literal and must not be recursed into.
-   */
-  static #unquote(v: JsonCodecValue): JsonCodecValue {
-    if (v === null || typeof v !== "object") {
-      return v;
-    } else if (Array.isArray(v)) {
-      const result = v.map(JsonCodecEngine.#unquote) as JsonCodecValue;
-      return Object.freeze(result);
-    } else if (
-      JsonCodecEngine.#isEncodedInstance(v) && Object.keys(v)[0] === "/quote"
-    ) {
-      return (v as Record<string, JsonCodecValue>)["/quote"]!;
-    } else {
-      const result = Object.fromEntries(
-        Object.entries(v).map(
-          ([k, val]) => [k, JsonCodecEngine.#unquote(val as JsonCodecValue)],
-        ),
-      ) as JsonCodecValue;
-      return Object.freeze(result);
-    }
-  }
-
-  /** Parses the JSON-text wire form, _without_ a tag prefix. */
-  static #parseWireText(jsonText: string): JsonCodecValue {
-    try {
-      return deepFreeze(JSON.parse(jsonText) as JsonCodecValue);
-    } catch (e) {
-      // The tag said this was ours and the text under it is not JSON, which
-      // is a refusal of the serialized form and settles against `lenient`
-      // like the tag check above it. Raised as this class's own refusal
-      // rather than passing `JSON.parse()`'s `SyntaxError` along, which
-      // nothing downstream recognizes.
-      const excerpt = (jsonText.length <= 50)
-        ? jsonText
-        : `${jsonText.slice(0, 50)}...`;
-      throw new ProblematicStateError(
-        "",
-        excerpt,
-        `Malformed JSON in an encoded \`FabricValue\` string: ${
-          backtickQuote(excerpt)
-        }`,
-        { cause: e },
-      );
-    }
   }
 }
